implement rr client creation/destruction
