+++ Source/AwsCommonRuntimeKit/mqtt/MqttRequestResponseClient.swift
@@ -1,0 +1,293 @@
+///  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+///  SPDX-License-Identifier: Apache-2.0.
+import AwsCMqtt
+import LibNative
+import Foundation
+
+/**
+ * The type of change to the state of a streaming operation subscription
+ */
+public enum SubscriptionStatusEventType {
+    /**
+     * The streaming operation is successfully subscribed to its topic (filter)
+     */
+    case established
+
+    /**
+     * The streaming operation has temporarily lost its subscription to its topic (filter)
+     */
+    case lost
+
+    /**
+     * The streaming operation has entered a terminal state where it has given up trying to subscribe
+     * to its topic (filter).  This is always due to user error (bad topic filter or IoT Core permission
+     * policy).
+     */
+    case halted
+}
+
+extension SubscriptionStatusEventType {
+    /// Returns the native representation of the Swift enum
+    var rawValue: aws_rr_streaming_subscription_event_type {
+        switch self {
+        case .established: return ARRSSET_SUBSCRIPTION_ESTABLISHED
+        case .lost: return ARRSSET_SUBSCRIPTION_LOST
+        case .halted: return ARRSSET_SUBSCRIPTION_HALTED
+        }
+    }
+
+    /// Initializes Swift enum from native representation
+    init(_ cEnum: aws_rr_streaming_subscription_event_type) {
+        switch cEnum {
+        case ARRSSET_SUBSCRIPTION_ESTABLISHED:
+            self = .established
+        case ARRSSET_SUBSCRIPTION_LOST:
+            self = .lost
+        case ARRSSET_SUBSCRIPTION_HALTED:
+            self = .halted
+        default:
+            fatalError("Unknown Susbscription Event Type")
+        }
+    }
+}
+
+/**
+ * An event that describes a change in subscription status for a streaming operation.
+ */
+ 
+public struct SubscriptionStatusEvent {
+    let event: SubscriptionStatusEventType
+    let errorCode: Int
+}
+
+/**
+ * An event that describes an incoming publish message received on a streaming operation.
+ *
+ * TODO: Igor have updated the events for IoT Command. Need update later
+ */
+public struct IncomingPublishEvent {
+
+    /// The payload of the publish message in a byte buffer format
+    let payload: Data
+
+    /// The topic associated with this PUBLISH packet.
+    let topic: String
+    
+    /// TODO: More options for IoT Command changes
+    /// ...
+}
+
+public typealias SubscriptionStatusEventHandler = (SubscriptionStatusEvent) async -> Void
+public typealias IncomingPublishEventHandler = (IncomingPublishEvent) async -> Void
+
+/// Encapsulates a response to an AWS IoT Core MQTT-based service request
+public struct MqttRequestResponseResponse {
+    let topic: String
+    let payload: Data
+}
+
+public struct ResponsePath {
+    let topic: String
+    let correlationTokenJsonPath: [String]
+}
+
+/// Generic configuration options for request response operation
+public struct RequestResponseOperationOptions: CStruct {
+    let subscriptionTopicFilters: [String]
+    let responsePaths: [ResponsePath]?
+    let topic: String
+    let payload: Data
+    let correlationToken: [String]?
+    
+    public init () {
+        // TODO: INIT THE MEMBERS
+        self.subscriptionTopicFilters = []
+        self.responsePaths = []
+        self.topic = ""
+        self.payload = Data()
+        self.correlationToken = nil
+    }
+
+    typealias RawType = aws_mqtt_request_operation_options
+    func withCStruct<Result>(_ body: (RawType) -> Result) -> Result {
+        // TODO: convert into aws_mqtt_request_operation_options
+        var options = aws_mqtt_request_operation_options()
+        return body(options)
+    }
+
+}
+
+/// Configuration options for streaming operations
+public struct StreamingOperationOptions: CStruct {
+    let subscriptionStatusEventHandler: SubscriptionStatusEventHandler
+    let incomingPublishEventHandler: IncomingPublishEventHandler
+    let topicFilter: String
+
+    public init () {
+        // TODO: INIT THE MEMBERS
+        self.subscriptionStatusEventHandler = {_ in return;}
+        self.incomingPublishEventHandler = {_ in return;}
+        self.topicFilter = ""
+    }
+
+    typealias RawType = aws_mqtt_streaming_operation_options
+    func withCStruct<Result>(_ body: (RawType) -> Result) -> Result {
+        // TODO: convert into aws_mqtt_request_operation_options
+        var options = aws_mqtt_streaming_operation_options()
+        return body(options)
+    }
+    
+}
+
+/**
+ * A streaming operation is automatically closed (and an MQTT unsubscribe triggered) when its
+ * destructor is invoked.
+ */
+public class StreamingOperation {
+    fileprivate var rawValue: OpaquePointer? // <aws_mqtt_rr_client_operation>?
+
+    public init () {
+        // TODO: INIT THE MEMBERS
+        self.rawValue = nil
+    }
+    
+    /**
+     * Opens a streaming operation by making the appropriate MQTT subscription with the broker.
+     */
+    public func open() {
+        // TODO:
+    }
+
+    deinit{
+        // TODO: close the oepration
+    }
+}
+
+public class MqttRequestResponseClientOptions: CStructWithUserData {
+
+    var maxRequestResponseSubscription: Int = 3
+    var maxStreamingSubscription: Int = 0
+    var operationTimeout: TimeInterval = 5
+    
+    init() {}
+    
+    /**
+     * Sets the maximum number of request-response subscriptions the client allows to be concurrently active
+     * at any one point in time.  When the client hits this threshold, requests will be delayed until
+     * earlier requests complete and release their subscriptions.  Each in-progress request will use either
+     * 1 or 2 MQTT subscriptions until completion.
+     *
+     * @param maxRequestResponseSubscriptions maximum number of concurrent subscriptions that the client
+     * will use for request-response operations
+     * @return reference to this
+     */
+    func withMaxRequestResponseSubscriptions(maxRequestResponseSubscriptions: Int) {
+        self.maxRequestResponseSubscription = maxRequestResponseSubscriptions
+    }
+
+    /// Sets the maximum number of concurrent streaming operation subscriptions that the client will allow.
+    /// Each "unique" (different topic filter) streaming operation will use 1 MQTT subscription.  When the
+    /// client hits this threshold, attempts to open new streaming operations will fail.
+    ///
+    /// @param maxStreamingSubscriptions maximum number of current subscriptions that the client will
+    /// use for streaming operations
+    func withMaxStreamingSubscriptions(maxStreamingSubscriptions: Int) {
+        self.maxStreamingSubscription = maxStreamingSubscriptions
+    }
+
+    /**
+     * Sets the timeout value, in seconds, for a request-response operation.  If a request is not complete
+     * by this time interval, the client will complete it as failed.  This time interval starts the instant
+     * the request is submitted to the client.
+     *
+     * @param operationTimeoutInSeconds request timeout in seconds
+     * @return reference to this
+     */
+    func withOperationTimeoutInSeconds(operationTimeoutInSeconds: TimeInterval) {
+        self.operationTimeout = operationTimeoutInSeconds
+    }
+
+    typealias RawType = aws_mqtt_request_response_client_options
+    func withCStruct<Result>(userData: UnsafeMutableRawPointer?, _ body: (aws_mqtt_request_response_client_options) -> Result) -> Result {
+        var options = aws_mqtt_request_response_client_options()
+        options.max_request_response_subscriptions = self.maxRequestResponseSubscription
+        options.max_streaming_subscriptions = self.maxStreamingSubscription
+        options.operation_timeout_seconds = try! self.operationTimeout.secondUInt32()
+        options.terminated_callback = MqttRRClientTerminationCallback
+        options.user_data = userData
+        return body(options)
+    }
+
+}
+
+internal func MqttRRClientTerminationCallback(_ userData: UnsafeMutableRawPointer?) {
+    // Termination callback. This is triggered when the native client is terminated.
+    // It is safe to release the request response client at this point.
+    // `takeRetainedValue()` would release the client reference. ONLY DO IT AFTER YOU NEED RELEASE THE CLIENT
+    _ = Unmanaged<MqttRequestResponseClientCore>.fromOpaque(userData!).takeRetainedValue()
+}
+
+internal class MqttRequestResponseClientCore {
+    fileprivate var rawValue: OpaquePointer? // <aws_mqtt_request_response_client>?
+    
+    internal init(mqttClient: Mqtt5Client, options: MqttRequestResponseClientOptions) throws {
+        guard let rawValue = (options.withCPointer(
+            userData: Unmanaged<MqttRequestResponseClientCore>.passRetained(self).toOpaque()) { optionsPointer in
+                return aws_mqtt_request_response_client_new_from_mqtt5_client(
+                    allocator, mqttClient.clientCore.rawValue, optionsPointer)
+            }) else {
+            // failed to create client, release the callback core
+            Unmanaged<MqttRequestResponseClientCore>.passUnretained(self).release()
+            throw CommonRunTimeError.crtError(.makeFromLastError())
+        }
+        self.rawValue = rawValue
+        
+    }
+    
+    /// submit a request responds operation, throws CRTError if the operation failed
+    public func submitRequest(operationOptions: RequestResponseOperationOptions) async throws -> MqttRequestResponseResponse {
+        // TODO: sumibt request
+        return MqttRequestResponseResponse(topic: "", payload: Data())
+    }
+    
+    /// create a stream operation, throws CRTError if the creation failed. You would need call open() on the operation to start the stream
+    public func createStream(streamOptions: StreamingOperationOptions) async throws -> StreamingOperation {
+        // TODO: create streamming operation
+        return StreamingOperation()
+    }
+    
+    public func close() {
+        aws_mqtt_request_response_client_release(self.rawValue)
+        self.rawValue = nil
+    }
+    
+}
+
+public class MqttRequestResponseClient {
+    fileprivate var clientCore: MqttRequestResponseClientCore
+    
+    public static func newFromMqtt5Client(
+        mqtt5Client: Mqtt5Client,
+        options: MqttRequestResponseClientOptions? = nil) throws -> MqttRequestResponseClient {
+            return try MqttRequestResponseClient(mqttClient: mqtt5Client, options: options ?? MqttRequestResponseClientOptions())
+    }
+    
+    internal init(mqttClient: Mqtt5Client, options: MqttRequestResponseClientOptions) throws {
+        clientCore = try MqttRequestResponseClientCore(mqttClient: mqttClient, options: options)
+    }
+    
+    /// submit a request responds operation, throws CRTError if the operation failed
+    public func submitRequest(operationOptions: RequestResponseOperationOptions) async throws -> MqttRequestResponseResponse {
+        return try await clientCore.submitRequest(operationOptions: operationOptions)
+    }
+    
+    /// create a stream operation, throws CRTError if the creation failed. You would need call open() on the operation to start the stream
+    public func createStream(streamOptions: StreamingOperationOptions) async throws -> StreamingOperation {
+        return try await clientCore.createStream(streamOptions: streamOptions)
+    }
+    
+    deinit{
+        self.clientCore.close()
+    }
+    
+}
